refactor(workspace): split the attach module along its seams
`engine-sync.ts` had grown to 1,436 lines holding five unrelated jobs. It is
now 916, and the other four live where they belong:

- `engine-types.ts`   vocabulary and pure predicates — no I/O, no ambient
                      state, so nothing here can be reordered against
                      anything else
- `engine-seams.ts`   ambient access and the single test seam
- `engine-probes.ts`  everything that asks the outside world a question
- `engine-config.ts`  the only path to configuration, always refreshed
- `engine-chain.ts`   per-project serialization

This commit moves code and moves nothing else. All 66 top-level blocks were
verified byte-identical against the previous head, modulo the `export`
keyword a cross-module reference requires — checked mechanically, not by
eye. The leaf helpers in particular are untouched: version comparison, the
version probe, the PATH lookup, HTTP bounding, the pin parser and the cache
bounds are the same lines they were, because the machine buys nothing there
and every touched line would be new surface.

`syncInternals` stays one flat object rather than becoming per-module seams.
It is the override surface consumers assign to, and splitting it would break
them for no gain. The public API is unchanged for the same reason: callers
import from `engine-sync` and should not have to learn where a symbol went.
